ポップアップが開かない環境のために UX 最適化、レイアウト整理
