+++ src/lib/pages/routes/popup.svelte
@@ -1,29 +1,18 @@
+<!--
+  @component
+  ポップアップに表示されるページ。モバイルでは、このページはポップアップではなく拡張機能専用ページで開かれる。ポップ
+  アップの内容を表示するのは冗長なので、直接履歴ページを開く。
+-->
 <script>
-  import { onMount } from 'svelte';
-  import PopupLayout from '$lib/pages/layouts/popup-layout.svelte';
-  import PopupHistory from '$lib/pages/popup/popup-history.svelte';
-  import PopupPlatformList from '$lib/pages/popup/popup-platform-list.svelte';
-  import { viewingHistory } from '$lib/services/history';
-
-  let currentPage = 'history';
-
-  const checkHash = () => {
-    currentPage = window.location.hash === '#/popup/platforms' ? 'platforms' : 'history';
-  };
-
-  onMount(() => {
-    checkHash();
-  });
+  import PopupContent from '$lib/pages/popup/popup-content.svelte';
+  import HistoryPage from '$lib/pages/routes/history.svelte';
+  import { isPopupOpen } from '$lib/services/runtime';
 </script>
 
-<svelte:window on:hashchange={() => checkHash()} />
-
-<PopupLayout>
-  {#if $viewingHistory}
-    {#if !$viewingHistory.length || currentPage === 'platforms'}
-      <PopupPlatformList />
-    {:else}
-      <PopupHistory />
-    {/if}
+{#await isPopupOpen() then isOpen}
+  {#if isOpen}
+    <PopupContent />
+  {:else}
+    <HistoryPage />
   {/if}
-</PopupLayout>
+{/await}
